clean(core): Remove un-used TypedUriIntoThingConverter (== UriIntoThingConverter)
